# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api for RosaFiesta a decoration enterprise management system",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "jeudrypp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
        "/articles": {
            "get": {
                "security": [
                    {
                        "StaticApiKey": []
                    }
                ],
                "description": "Get all articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get all Articles",
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Article"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "StaticApiKey": []
                    }
                ],
                "description": "Creates a new article with variants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Creates Article",
                "parameters": [
                    {
                        "description": "Article creation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateProductPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created article",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/articles/{articleId}": {
            "get": {
                "security": [
                    {
                        "StaticApiKey": []
                    }
                ],
                "description": "Get an article by its ID with variants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "StaticApiKey": []
                    }
                ],
                "description": "Delete an article by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Delete Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/articles/{id}": {
            "put": {
                "security": [
                    {
                        "StaticApiKey": []
                    }
                ],
                "description": "Update an article basic info (not variants yet)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Update Article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Article update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateProductPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated article",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/authentication/refresh": {
            "post": {
                "description": "Get a new access token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Refresh an access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New tokens",
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/authentication/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RegisterUserPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserWithToken"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/authentication/token": {
            "post": {
                "description": "Create a new token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Create a new token",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateUserTokenPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateArticleVariantPayload": {
            "type": "object",
            "required": [
                "name",
                "rental_price",
                "sku"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateArticleDimensionPayload"
                    }
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "rental_price": {
                    "type": "number",
                    "minimum": 0
                },
                "replacement_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "sale_price": {
                    "type": "number",
                    "minimum": 0
                },
                "sku": {
                    "type": "string",
                    "maxLength": 255
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "dtos.CreateProductPayload": {
            "type": "object",
            "required": [
                "name_template",
                "type",
                "variants"
            ],
            "properties": {
                "category_id": {
                    "description": "string to parsing uuid later",
                    "type": "string"
                },
                "description_template": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_template": {
                    "type": "string",
                    "maxLength": 100
                },
                "type": {
                    "enum": [
                        "Rental",
                        "Sale"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ArticleType"
                        }
                    ]
                },
                "variants": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dtos.CreateArticleVariantPayload"
                    }
                }
            }
        },
        "dtos.CreateUserTokenPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3
                }
            }
        },
        "dtos.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "accessTokenExpirationTimestamp": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dtos.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "dtos.RegisterUserPayload": {
            "type": "object",
            "required": [
                "confirm_password",
                "email",
                "first_name",
                "last_name",
                "password",
                "username"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "dtos.UpdateProductPayload": {
            "type": "object",
            "required": [
                "name_template",
                "type"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description_template": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_template": {
                    "type": "string",
                    "maxLength": 100
                },
                "type": {
                    "enum": [
                        "Rental",
                        "Sale"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ArticleType"
                        }
                    ]
                }
            }
        },
        "dtos.UserWithToken": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "born_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "role_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "models.Article": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "For eager loading",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ]
                },
                "category_id": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "description_template": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name_template": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.ArticleType"
                },
                "updated": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ArticleVariant"
                    }
                }
            }
        },
...
        "models.ArticleType": {
            "type": "string",
            "enum": [
                "Rental",
                "Sale"
            ],
            "x-enum-varnames": [
                "ArticleTypeRental",
                "ArticleTypeSale"
            ]
        },
        "models.ArticleVariant": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "string"
                },
                "attributes": {
                    "description": "Relations",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ArticleDimension"
                    }
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "rental_price": {
                    "type": "number"
                },
                "replacement_cost": {
                    "type": "number"
                },
                "sale_price": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Provide your API key to access this API",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "StaticApiKey": {
            "description": "Provide your Static API key to access this API",
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        }
    }
}